# perf: avoid exceptions in extract headers